kubeconfig: Use config rather than environment
We used set an environment variable to automatically `--namespace` and
`--context` prepend commands based on the kubectl context commands that
were ran. Now, we actually construct a kubeconfig file with the proper
context/namespace, and switch `KUBECONFIG` to look at this instead,
which means we don't have to alter the command line.

TODO: We need to be able to purge the files sometimes?